+++ src/languages/hive/hive.functions.ts
@@ -1,4 +1,4 @@
-import type { FunctionSignature } from '../../completion/functionSignatures'
+import type { FunctionSignature } from '../../completion'
 
 export const functions: string[] = [
     // https://cwiki.apache.org/confluence/display/Hive/LanguageManual+UDF