+++ src/languages/spark/spark.functions.ts
@@ -1,4 +1,4 @@
-import type { FunctionSignature } from '../../completion/functionSignatures'
+import type { FunctionSignature } from '../../completion'
 
 export const functions: string[] = [
     // http://spark.apache.org/docs/latest/sql-ref-functions.html